pyproject.toml
[tool.poetry]
name = "drfprogect"
version = "0.1.0"
description = ""
authors = ["Andrew_Polik <polik.24@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
